rename file to snake case
