+++ src/spetlr/etl/loaders/__init__.py
@@ -1,4 +1,4 @@
 from .simple_loader import SimpleLoader  # noqa: F401
 from .simple_sql_loader import SimpleSqlServerLoader  # noqa: F401
+from .upsert_loader_streaming import UpsertLoaderStreaming  # noqa: F401
 from .UpsertLoader import UpsertLoader  # noqa: F401
-from .UpsertLoaderStreaming import UpsertLoaderStreaming  # noqa: F401
